requirements.txt
langchain
langchain-groq
langchain-community
youtube_transcript_api
pytube
pypdf
pandas
python-dotenv
